evaluate darcs-to-git
ugh, that was quick. i expected it to pass most of my testcases, but it
fails before the interesting stuff. :)

long: i wanted to see how does it deal with the darcs chan --xml -s
output (where tailor fails), but it fails with trivial history as well
if the repo contains funny author names.