darcs-fast-import: fix timezone processing
I observed git-to-darcs resulting in a darcs patch with a timestamp that was
7 hours later than it should have been (I live in PDT, -0700). I think
get_date() was wrong: it generated a UTC-based string for darcs, but then
appended the committer's original timezone. Darcs thought this meant it
should treat the (UTC) string as localtime-in-that-timezone, and applied an
offset it shouldn't have.

If we only cared about preserving the committer time, then it would be
appropriate to pass a UTC string with no timezone into darcs. But we're
trying to record two things: the absolute commit time, and the localtime in
the committers timezone (i.e. the absolute commit time and the timezone in
which it occurred).

My new version of get_date() tries to accomplish this by parsing the
timezone-offset we get from git (well, from the git-fast-export -format
stream), turn it into a number of seconds, then add that number to the
absolute timestamp we got from the stream. Then it uses time.gmtime() to
render that (effectively getting a localtime string from the committer's
timezone), and then appends the same timezone string as before. Darcs will
then effectively undo this combination, to derive the absolute
seconds-since-epoch and also record the localtime string.

With this patch, my darcs timestamps appear the same as my git ones.